# docs/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /public/verify:
    post:
      tags:
      - public-controller
      operationId: verifyEmail
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /public/signup:
    post:
      tags:
      - public-controller
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /public/login:
    post:
      tags:
      - public-controller
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /user-panel/logout:
    get:
      tags:
      - user-controller
      operationId: logout
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /user-panel/find-user/{userName}:
    get:
      tags:
      - user-controller
      operationId: findUser
      parameters:
      - name: userName
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /user-panel/add-balance:
    get:
      tags:
      - user-controller
      operationId: addBalance
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      - name: balance
        in: query
        required: true
        schema:
          type: number
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /stock/{id}:
    get:
      tags:
      - stock-controller
      operationId: getStockById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /stock/search:
    get:
      tags:
      - stock-controller
      operationId: searchStocks
      parameters:
      - name: name
        in: query
        required: false
        schema:
          type: string
      - name: sector
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /public/verification:
    get:
      tags:
      - public-controller
      operationId: verification
      parameters:
      - name: token
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /user-panel/delete:
    delete:
      tags:
      - user-controller
      operationId: deleteUser
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
components:
  schemas:
    AuthRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userName:
          type: string
        password:
          type: string
        email:
          type: string
        role:
          type: array
          items:
            type: string
        createadAt:
          type: string
          format: date-time
        lastActive:
          type: string
          format: date-time
        balance:
          type: number
        country:
          type: string
        status:
          type: string
        verificationToken:
          type: string
        resetToken:
          type: string
        verified:
          type: boolean
...